LayoutTests/imported/w3c/web-platform-tests/css/css-tables/tentative/colgroup-col-expected.txt

Colgroup col widths

colgroup's purpose is to "group" cols
one can think of colgroup as always generating max(colgroup.span,1) cols, unless it has <col> children.
colgroup css_width specifies width of generated cols.
col.css_width can override colgroup width
colgroups with spans
cg1	cg2	cg2	cg2	50px
cols with spans
col1	col2	col2	col2	50px
colgroups with cols
cg	col1	col2	50px
colgroups with cols with spans
Legacy fails col with span inside colgroup test.
cg1	col1	col1	cg2	cg2	50px
0% cols are ignored
01234	
01234	

PASS table 1
PASS table 2
PASS table 3
FAIL table 4 assert_equals:
<table>
  <colgroup span="3" style="width:100px">
    <col>
    <col style="width:50px" span="2">
  </colgroup>
  <colgroup style="width:66px">
    <col span="2">
  </colgroup>
  <tbody><tr><td data-expected-width="100">cg1</td>
  <td data-expected-width="50">col1</td>
  <td data-expected-width="50">col1</td>
  <td data-expected-width="66">cg2</td>
  <td data-expected-width="66">cg2</td>
  <td data-expected-width="75"><div style="width:75px">50px</div></td>
</tr></tbody></table>
width expected 66 but got 23
PASS table 5
PASS table 6
